COMPLETE but need Refactor

+++ src/main/java/com/dongnv/courseregistrationmanagement/service/EnrollmentService.java
@@ -1,12 +1,19 @@
 package com.dongnv.courseregistrationmanagement.service;
 
+import com.dongnv.courseregistrationmanagement.dto.PageResponse;
+import com.dongnv.courseregistrationmanagement.dto.mapper.CourseMapper;
+import com.dongnv.courseregistrationmanagement.dto.mapper.UserMapper;
 import com.dongnv.courseregistrationmanagement.dto.request.EnrollCourseRequest;
 import com.dongnv.courseregistrationmanagement.dto.request.UnenrollCourseRequest;
 import com.dongnv.courseregistrationmanagement.dto.response.EnrollCourseResponse;
+import com.dongnv.courseregistrationmanagement.dto.response.StudentInCourseResponse;
+import com.dongnv.courseregistrationmanagement.dto.response.UserEnrollmentResponse;
+import com.dongnv.courseregistrationmanagement.dto.response.UserResponse;
 import com.dongnv.courseregistrationmanagement.exception.AppException;
 import com.dongnv.courseregistrationmanagement.exception.ErrorCode;
 import com.dongnv.courseregistrationmanagement.model.Course;
 import com.dongnv.courseregistrationmanagement.model.Enrollment;
+import com.dongnv.courseregistrationmanagement.model.User;
 import com.dongnv.courseregistrationmanagement.repository.CourseRepository;
 import com.dongnv.courseregistrationmanagement.repository.EnrollmentRepository;
 import lombok.AccessLevel;
@@ -14,6 +21,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.jdbc.datasource.DataSourceTransactionManager;
 import org.springframework.orm.jpa.JpaTransactionManager;
 import org.springframework.security.core.Authentication;
@@ -23,6 +32,7 @@ import org.springframework.transaction.PlatformTransactionManager;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDate;
+import java.util.List;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.locks.Lock;
 import java.util.concurrent.locks.ReentrantLock;
@@ -35,6 +45,7 @@ public class EnrollmentService {
     EnrollmentRepository enrollmentRepository;
     CourseRepository courseRepository;
     ConcurrentHashMap<Long, Lock> locks = new ConcurrentHashMap<>();
+    CourseMapper courseMapper;
     
     @Transactional(transactionManager = "primaryTransactionManager")
     public EnrollCourseResponse enrollCourse(EnrollCourseRequest request) {
@@ -47,9 +58,7 @@ public class EnrollmentService {
 
         // Lock block have race conditional
         Lock lock = locks.computeIfAbsent(course.getId(), key -> new ReentrantLock());
-        lock.lock();
-        saveEnrollment(course, enrollment);
-        lock.unlock();
+        saveEnrollment(course, enrollment, lock);
 
         return EnrollCourseResponse.builder()
                 .status(true)
@@ -57,15 +66,40 @@ public class EnrollmentService {
                 .build();
     }
 
-    @Transactional
+    @Transactional(transactionManager = "primaryTransactionManager")
     public void unenrollCourse(UnenrollCourseRequest request) {
         Course course = getCourseAndCheckCondition(request.getCourseId());
         Long id = getCurrentUserId();
 
+        if (LocalDate.now().isAfter(course.getStartDate()))
+            throw new AppException(ErrorCode.COURSE_IS_STARTED);
         enrollmentRepository.deleteEnrollmentByUserIdAndCourseId(id, request.getCourseId());
 
         course.setCurrentEnrollments(course.getCurrentEnrollments() - 1);
         courseRepository.save(course);
+    }
+
+    public boolean isEnroll(Long courseId) {
+        Long id = getCurrentUserId();
+        return enrollmentRepository.existsByUserIdAndCourseId(id, courseId);
+    }
+
+    public StudentInCourseResponse getStudentInCourseById(Long courseId, int page, int size) {
+        Course course = courseRepository.findById(courseId).orElseThrow(() -> new AppException(ErrorCode.COURSE_NOT_FOUND));
+        Page<UserEnrollmentResponse> userPage = enrollmentRepository.getUsersByCourseId(courseId, PageRequest.of(page - 1, size));
+
+        PageResponse<UserEnrollmentResponse> userPageResponse = PageResponse.<UserEnrollmentResponse>builder()
+                .currentPage(page)
+                .totalPages(userPage.getTotalPages())
+                .pageSize(userPage.getSize())
+                .totalElements(userPage.getTotalElements())
+                .data(userPage.getContent())
+                .build();
+
+        return StudentInCourseResponse.builder()
+                .course(courseMapper.toCourseResponse(course))
+                .users(userPageResponse)
+                .build();
     }
 
 
@@ -85,8 +119,9 @@ public class EnrollmentService {
     }
 
 
-    private void saveEnrollment(Course course, Enrollment enrollment) {
+    private void saveEnrollment(Course course, Enrollment enrollment, Lock lock) {
         try {
+            lock.lock();
             // Check course is full
             if (!courseRepository.currentEnrollmentsLessThanMaxEnrollmentsById(course.getId()))
                 throw new AppException(ErrorCode.COURSE_IS_FULL);
@@ -96,6 +131,8 @@ public class EnrollmentService {
             courseRepository.save(course);
         } catch (DataIntegrityViolationException e) {
             throw new AppException(ErrorCode.ENROLLMENT_EXISTED);
+        } finally {
+            lock.unlock();
         }
     }
 }
